TVG-116 Move call to toggle a SearchItem's status to the component
