+++ frontend/src/components/SearchItem/index.tsx
@@ -1,4 +1,4 @@
-import { useContext, useState } from "react";
+import { Dispatch, SetStateAction, useContext, useState } from "react";
 import { App, Button, Dropdown, Popconfirm, Table, Tag, Typography } from "antd";
 import type { MenuProps, TableColumnsType } from "antd";
 import { DeleteFilled, EditOutlined, PoweroffOutlined } from "@ant-design/icons";
@@ -11,9 +11,10 @@ import {
     addSearchCriteria,
     deleteSearchCriteria,
     editSearchCriteria,
+    toggleStatus,
 } from "../../requests/search-items";
 import { handleErrorResponse } from "../../utils";
-import type { FormMode, SearchItem, SearchItemPayload } from "../../utils/types";
+import type { FormMode, SearchItem, SearchItemPayload, ShowData } from "../../utils/types";
 import "./SearchItem.css";
 
 
@@ -21,10 +22,10 @@ interface SearchItemProps {
     searchItem: SearchItem
     show: string
     displayActions?: boolean;
-    toggleStatus: () => Promise<void>;
+    setShowData: Dispatch<SetStateAction<ShowData>>;
 }
 
-const SearchItem = ({ searchItem, show, displayActions, toggleStatus }: SearchItemProps) => {
+const SearchItem = ({ searchItem, show, displayActions, setShowData }: SearchItemProps) => {
     const [openModal, setOpenModal] = useState(false);
     const [formMode, setFormMode] = useState<FormMode>(null);
 
@@ -65,6 +66,30 @@ const SearchItem = ({ searchItem, show, displayActions, toggleStatus }: SearchIt
                 message: `Unable to delete the search criteria for ${show}`,
                 description: message,
                 duration: 8
+            });
+        }
+    };
+
+    const toggleSearch = async () => {
+        const newStatus = searchItem.search_active ? false : true;
+        try {
+            const response = await toggleStatus(searchItem.id, newStatus);
+            console.log("response", response)
+            setShowData(current => ({ ...current, search_item: response }));
+            updateShowContext(show, "search_item", response);
+            notification.success({
+                message: "Success!",
+                description: `The search status for ${show} has been updated.`,
+            });
+        }
+        catch(error) {
+            let message: string = error?.message;
+            if (error?.response) {
+                message = handleErrorResponse(error, "update the search status for this show");
+            }
+            notification.error({
+                message: `There was a problem updating ${show}!`,
+                description: message,
             });
         }
     };
@@ -144,7 +169,7 @@ const SearchItem = ({ searchItem, show, displayActions, toggleStatus }: SearchIt
             icon: <PoweroffOutlined />,
             key: "toggle-active",
             label: searchItem?.search_active ? "Deactivate" : "Activate",
-            onClick: toggleStatus,
+            onClick: toggleSearch,
         },
         {
             icon: <EditOutlined />,
